refactor: split domain::repository
