+++ api/domain/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod external;
 pub mod model;
 pub mod repository;
